+++ commons/dimas-commands/src/lib.rs
@@ -6,29 +6,29 @@
 use std::collections::HashMap;
 // region:		--- modules
 use derivative::Derivative;
+use dimas_com::Communicator;
+use dimas_config::Config;
 use itertools::Itertools;
 use std::fmt::Display;
 use std::time::Duration;
-use zenoh::config::{Config, WhatAmI};
+use zenoh::config::{Locator, WhatAmI};
 use zenoh::prelude::sync::*;
 // endregion:	--- modules
 
-// region:		--- DimasEntity
-/// A `DiMAS` entity
+// region:		--- ScoutingEntity
+/// A `Zenoh` entity
 #[derive(Derivative)]
 #[derivative(Clone, Debug, PartialEq, PartialOrd, Eq, Ord)]
-pub struct DimasEntity {
-	name: String,
+pub struct ScoutingEntity {
 	zid: String,
-	#[derivative(PartialOrd="ignore", Ord="ignore")]
 	kind: String,
-	#[derivative(PartialOrd="ignore", Ord="ignore")]
+	#[derivative(PartialOrd = "ignore", Ord = "ignore")]
 	locators: Vec<Locator>,
 }
 
-impl Display for DimasEntity {
+impl Display for ScoutingEntity {
 	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		f.debug_struct("DimasEntity")
+		f.debug_struct("ScoutingEntity")
 			.field("zid", &self.zid)
 			.field("kind", &self.kind)
 			.field("locators", &self.locators)
@@ -36,36 +36,29 @@ impl Display for DimasEntity {
 	}
 }
 
-impl DimasEntity {
-	/// Fetch the list of reachable entities sorted by name of entity
+impl ScoutingEntity {
+	/// Sout for `DiMAS` entities, sorted by zid of entity
 	/// # Panics
 	/// if something goes wrong
 	#[must_use]
-	pub fn fetch(config: &Config) -> Vec<Self> {
+	pub fn scout(config: &Config) -> Vec<Self> {
 		let mut map: HashMap<String, Self> = HashMap::new();
 		let what = WhatAmI::Router | WhatAmI::Peer | WhatAmI::Client;
-		let receiver = zenoh::scout(what, config.clone())
+		let receiver = zenoh::scout(what, config.zenoh_config())
 			.res()
 			.expect("scouting failed");
 
 		while let Ok(hello) = receiver.recv_timeout(Duration::from_millis(250)) {
 			let zid = hello.zid.to_string();
 			map.entry(zid.clone()).or_insert(Self {
-				name: zid.to_string(),
 				zid,
 				kind: hello.whatami.to_string(),
 				locators: hello.locators,
 			});
 		}
 		let result: Vec<Self> = map.values().sorted().cloned().collect();
-		
-		result
-	}
 
-	/// Get the Name
-	#[must_use]
-	pub fn name(&self) -> &str {
-		&self.name
+		result
 	}
 
 	/// Get the Zenoh ID
@@ -74,10 +67,56 @@ impl DimasEntity {
 		&self.zid
 	}
 
-	/// Get the Kind 
+	/// Get the Kind
 	#[must_use]
 	pub fn kind(&self) -> &str {
 		&self.kind
 	}
+
+	/// Get the Locators
+	#[must_use]
+	pub const fn locators(&self) -> &Vec<Locator> {
+		&self.locators
+	}
 }
-// endregion:	--- DimasEntity
+// endregion:	--- ScoutingEntity
+
+// region:		--- about_list
+/// Fetch a list of about messages from all reachable `DiMAS` entities
+/// # Panics
+#[must_use]
+pub fn about_list(com: &Communicator) -> Vec<String> {
+	let map: HashMap<String, String> = HashMap::new();
+
+	let selector = String::from("**/about");
+
+	// fetch about from all entities
+	let replies = com
+		.session
+		.get(&selector)
+		.consolidation(ConsolidationMode::None)
+		.target(QueryTarget::All)
+		.allowed_destination(Locality::Any)
+		.timeout(Duration::from_millis(1000))
+		.res()
+		.expect("failed to create 'Receiver'");
+
+	while let Ok(reply) = replies.recv() {
+		match reply.sample {
+			Ok(sample) => println!(
+				">> Received ('{}': '{}')",
+				sample.key_expr.as_str(),
+				sample.value,
+			),
+			Err(err) => println!(
+				">> Received (ERROR: '{}')",
+				String::try_from(&err).expect("snh")
+			),
+		}
+	}
+
+	let result: Vec<String> = map.values().sorted().cloned().collect();
+
+	result
+}
+// endregion:	--- about_list