+++ dimasctl/src/main.rs
@@ -4,6 +4,7 @@
 
 // region:		--- modules
 use clap::{Parser, Subcommand};
+use dimas_com::Communicator;
 use dimas_config::Config;
 // endregion:	--- modules
 
@@ -19,28 +20,36 @@ struct DimasctlArgs {
 // region:		--- Commands
 #[derive(Debug, Subcommand)]
 enum DimasctlCommand {
-	/// List all agents
+	/// List running `DiMAS` entities
 	List,
+	/// Scout for `Zenoh` entities
+	Scout,
 }
 // endregion:	--- Commands
 
 fn main() {
 	let args = DimasctlArgs::parse();
+	let config = Config::default();
 
 	match &args.command {
 		DimasctlCommand::List => {
-			let config = Config::default().zenoh_config();
-			// no changes with these config modifications!
-			//let mut config = Config::default().zenoh_config();
-			//let mut scouting_cfg = config.scouting().clone();
-			//let mut gossip_cfg = scouting_cfg.gossip().clone();
-			//gossip_cfg.set_enabled(Some(true)).expect("snh");
-			//scouting_cfg.set_gossip(gossip_cfg).expect("snh");
-			//config.set_scouting(scouting_cfg).expect("snh");
-			println!("List of available agents:");
-			println!("{:32}  {:6}  {:5}", "ZenohId", "Kind", "Name");
-			for item in dimas_commands::DimasEntity::fetch(&config) {
-				println!("{:32}  {:6}  {}", item.zid(), item.kind(), item.name());
+			let com = Communicator::new(config).expect("failed to create 'Communicator'");
+			println!("List of running DiMAS entities:");
+			println!("{:32}  {:6}  Name", "ZenohId", "Kind");
+			for item in dimas_commands::about_list(&com) {
+				println!("{item}");
+			}
+		}
+		DimasctlCommand::Scout => {
+			println!("List of scouted Zenoh entities:");
+			println!("{:32}  {:6}  Locators", "ZenohId", "Kind");
+			for item in dimas_commands::ScoutingEntity::scout(&config) {
+				println!(
+					"{:32}  {:6}  {:?}",
+					item.zid(),
+					item.kind(),
+					item.locators()
+				);
 			}
 		}
 	}